requirements.txt
Js2Py==0.74
Pillow==9.3.0
requests==2.28.1
streamlit==1.14.0
streamlit_lottie==0.0.3
